+++ code/module/head/main/ble.c
@@ -327,7 +327,7 @@ static void gatts_profile_a_event_handler(esp_gatts_cb_event_t event, esp_gatt_i
     }
     case ESP_GATTS_WRITE_EVT: {
         ESP_LOGI(GATTS_TAG, "GATT_WRITE_EVT, conn_id %d, trans_id %d, handle %d",
-                 param->write.conn_id, param->write.trans_id, param->write.handle);
+                 param->write.conn_id, (int) param->write.trans_id, param->write.handle);
 
         if (!param->write.is_prep) {
             ESP_LOGI(GATTS_TAG, "GATT_WRITE_EVT, value len %d, value:", param->write.len);
